fix: cards page
