+++ src/common/selectors/Selectors.ts
@@ -37,3 +37,4 @@ export const resetRange = (state: RootStateType) => state.packs.resetRange
 //cards
 export const cardsSelector = (state: RootStateType) => state.cards.cards
 export const cardsPackName = (state: RootStateType) => state.cards.packName
+export const cardCreatorId = (state: RootStateType) => state.cards.creatorId
