+++ src/common/style/mixins.scss
@@ -58,6 +58,23 @@
   line-height: 20px;
 }
 
+@mixin navigateLink() {
+
+  @include subtitle(500);
+
+  display: flex;
+  column-gap: 12px;
+  align-items: center;
+  align-self: flex-start;
+
+  color: inherit;
+  text-decoration: none;
+
+  &:hover {
+    color: $primary-color
+  }
+}
+
 //other
 @mixin popup() {
   display: flex;
