--- src/pages/cards/cardsList/CardsList.module.scss
@@ -1,13 +1,0 @@
-.buttonsContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between
-}
-
-.searchContainer {
-  margin: 20px 0;
-}
-
-.search {
-  width: 100%;
-}
