+++ src/pages/cards/cardsList/CardsList.tsx
@@ -1,15 +1,7 @@
 import React, { FC } from 'react'
 
-import { useDispatch } from 'react-redux'
-import { useNavigate, useParams } from 'react-router-dom'
-
-import s from './CardsList.module.scss'
-
-import { setIsLoading, toggleModal } from 'app/appSlice'
-import { Button } from 'common/components/button/Button'
-import { AddCardModal } from 'common/components/modals/AddCardModal'
 import { useAppSelector } from 'common/hooks/useAppSelector'
-import { Search } from 'common/modules/search/Search'
+import { cardCreatorId, myIdSelector } from 'common/selectors/Selectors'
 import list from 'common/style/List.module.scss'
 import { formatDate } from 'common/utils/formatDate'
 import { CardType } from 'pages/cards/cardsApi'
@@ -21,66 +13,37 @@ type CardsListType = {
 }
 
 export const CardsList: FC<CardsListType> = ({ cards }) => {
-  const myId = useAppSelector(state => state.auth.id)
-  const packCreatorId = useAppSelector(state => state.cards.creatorId)
+  const myId = useAppSelector(myIdSelector)
+  const packCreatorId = useAppSelector(cardCreatorId)
   const myPack = myId === packCreatorId
 
-  const navigate = useNavigate()
-  const dispatch = useDispatch()
-  const { id } = useParams()
-
-  const handelLearnPack = () => {
-    dispatch(setIsLoading(true))
-    navigate(`/cards/${id}/learn`)
-  }
-
   return (
-    <>
-      <div className={s.buttonsContainer}>
-        <Button styleType="primary" onClick={handelLearnPack}>
-          learn pack
-        </Button>
+    <table className={list.table}>
+      <thead>
+        <tr>
+          <th className={list.tableTitle}>Question</th>
+          <th className={list.tableTitle}>Answer</th>
+          <th className={list.tableTitle}>Last Updated</th>
+          <th className={list.tableTitle}>Grade</th>
+          {myPack && <th className={list.tableTitle}>Actions</th>}
+        </tr>
+      </thead>
+      <tbody className={list.tableBody}>
+        {cards.map(c => {
+          const dateUpdate = formatDate(c.updated)
 
-        {myPack && (
-          <Button onClick={() => dispatch(toggleModal(true))} styleType={'primary'}>
-            Add New Card
-          </Button>
-        )}
-      </div>
-
-      <AddCardModal />
-
-      <div className={s.searchContainer}>
-        <Search class={s.search} />
-      </div>
-
-      <table className={list.table}>
-        <thead>
-          <tr>
-            <th className={list.tableTitle}>Question</th>
-            <th className={list.tableTitle}>Answer</th>
-            <th className={list.tableTitle}>Last Updated</th>
-            <th className={list.tableTitle}>Grade</th>
-            {myPack && <th className={list.tableTitle}>Actions</th>}
-          </tr>
-        </thead>
-        <tbody className={list.tableBody}>
-          {cards.map(c => {
-            const dateUpdate = formatDate(c.updated)
-
-            return (
-              <Card
-                key={c._id}
-                question={c.question}
-                answer={c.answer}
-                update={dateUpdate}
-                grade={Number(c.grade.toFixed(2))}
-                actions={myPack && <CardActions cardId={c._id} question={c.question} answer={c.answer} />}
-              />
-            )
-          })}
-        </tbody>
-      </table>
-    </>
+          return (
+            <Card
+              key={c._id}
+              question={c.question}
+              answer={c.answer}
+              update={dateUpdate}
+              grade={Number(c.grade.toFixed(2))}
+              actions={myPack && <CardActions cardId={c._id} question={c.question} answer={c.answer} />}
+            />
+          )
+        })}
+      </tbody>
+    </table>
   )
 }
