+++ src/pages/cards/cardsList/learnCard/LearnCard.module.scss
@@ -47,17 +47,5 @@
 
 .link {
 
-  @include subtitle(500);
-
-  display: flex;
-  column-gap: 12px;
-  align-items: center;
-  align-self: flex-start;
-
-  color: inherit;
-  text-decoration: none;
-
-  &:hover {
-    color: $primary-color
-  }
+  @include navigateLink();
 }
